move add node util functions to flowUtils
